refactoring

+++ app/controllers/api/v1/negotiation_principals_controller.rb
@@ -6,23 +6,13 @@ class Api::V1::NegotiationPrincipalsController < ApplicationController
     render json: negotiation_principals
   end
 
-  # def show
-  #   negotiation_principal = NegotiationPrincipal.where(negotiation_id: params['negotiation_id'],
-  #     principal_id: params['principal_id']).first
-  #   render json: {
-  #     neg_id: negotiation_principal.negotiation_id,
-  #     rating: negotiation_principal.rating,
-  #     traded: negotiation_principal.traded
-  #   }
-  # end
 
   def update
     neg_prin = NegotiationPrincipal.find(params['id'])
     neg_prin.update(update_params)
-    ioi = Ioi.where(principal_id: neg_prin.principal_id, side: neg_prin.side, active: false)
-    ioi.destroy
+    ioi = Ioi.where(principal_id: neg_prin.principal_id, side: neg_prin.side, active: false)[0]
+    ioi.destroy unless !ioi
     # Deletes IOI so, algo controller can search active and non-active IOIs for matches.
-    binding.pry
     render json: neg_prin
   end
